fix: imports, renamed prop
