+++ src/components/shared/Extensions/Button/types.ts
@@ -14,5 +14,5 @@ export interface ButtonProps extends Omit<ButtonHTMLAttributes<HTMLButtonElement
   textValues?: SimpleMessageValues;
   ariaLabel?: MessageDescriptor | string;
   isFullSize?: boolean;
-  fullRounded?: boolean;
+  isFullRounded?: boolean;
 }
